Fix delayed task restart after errors
Previously, when scheduling a new write, I called

  Process.send_after(__MODULE__, (...))

But this caused problems when a process crashes: __MODULE__ always
exists, so the scheduled call would not disappear. When a new process is
started by the Supervisor, that also scheduled the same call, causing an
additional task to be scheduled for each process crash.

The correct way to do it is to _not_ send it to __MODULE__ but to
self().

When a process crashes, the scheduled call has nowhere to go and is
ignored.

Then, when the process is restarted, it will schedule the call again,
thus ensuring that only one call is ever being scheduled at the same
time.