Never stop the supervisor if children crash.
Stopping the supervisor stops the entire application, which needs manual
intervention to restart again.

Since it has never been necessary stop it, and the fact that it happened
one time due to a temporary issue in DST handing, losing 3 days of data,
I think it makes sense to just keep it running.